pyproject.toml
[tool.poetry]
name = "dr-sven"
version = "0.1.0-alpha"
description = "A naive serverless data health checker"
authors = ["MeStrak <mestrak@github.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
pandas = "^1.1.0"
toml = "^0.10.1"
awswrangler = "^1.8.1"
dateparser = "^0.7.6"
nanoid = "^2.0.0"
python-slugify = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
